Assign MRs instead of requesting review
Change the dependency upgrade script so that it will assign the merge
request to the chosen person instead of just adding them as a reviewer.
This makes it more clear that this person is the DRI for the MR and is
responsible for seeing it merged. "Request review" only implies
responsibility for reviewing the code.